log: avoid code duplication, add missing freelog const, add mood
