+++ commands/LOGHELP.md
@@ -3,3 +3,5 @@
 **To log in a free format:** Type `+freelog` then write your adaptation log in the private message the bot sent you.
 
 **To log using the bot's help:** Type `+log` and follow the bot instructions sent to you in a private message on Discord.
+
+**To post an EEG log:** Type `+logeeg` and send the bot your eeg screenshot in a private missage.
