+++ server/commands/logeeg.dm.js
@@ -1,12 +1,9 @@
 const config = require("../../config.json");
 const Discord = require('discord.js');
-const _ = require("lodash");
+require('./log.tools.js')();
 
 let currentUsers = [];
 
-const timeout = 3600;
-const timeoutMessage = 'Timeout: the process was aborted.';
-const abortMessage = 'Process was aborted.';
 const logEegCMD = 'logeeg';
 const logsChannelName = 'adaptation_logs';
 
@@ -32,7 +29,7 @@ async function logEeg(message, dry = false) {
 
     let botMessage;
     try {
-      botMessage = await message.author.send('Send your EEG here. Its optional comment will also be posted. If you want to abort, wait an hour or answer here with `abort`.');
+      botMessage = await message.author.send('You can send your EEG. Its comment will also be posted (optional). If you want to abort, wait an hour or answer with `abort` or `X`.');
     }
     catch (err) {
       console.warn(`WARN\t: Couldn't send message to ${message.author.username}: ${err}`);
@@ -58,7 +55,7 @@ async function logEeg(message, dry = false) {
     }
 
     if (!qSleepTracker.attachment) {
-      message.author.send("No EEG attachment found, please try again.");
+      message.author.send("No EEG attachment found, please use the command again and then attach your EEG file.");
       currentUsers.splice(currentUsers.indexOf(message.author.id), 1);
       return true;
     }
@@ -75,48 +72,4 @@ async function logEeg(message, dry = false) {
     message.author.send('You must join the Polyphasic Sleeping server if you want to post adaptation logs.');
   }
   return true;
-}
-
-async function collectFromUser(author, channel, step) {
-  try {
-    let collected = await channel.awaitMessages(x => x.author.id === author.id, { maxMatches: 1, time: timeout * 1000, errors: ['time'] });
-    if (collected.first().content.toLowerCase() === 'abort') {
-      author.send('Aborted.');
-      return null;
-    }
-    return collected.first();
-  }
-  catch (e) {
-    console.log("WARN\t: ", `Timeout waiting for answer from ${author.username} during step ${step.name}`);
-    author.send(timeoutMessage);
-    return null;
-  }
-}
-
-async function processqSleepTracker(message, qSleepTracker) {
-  if (!(collected = await collectFromUser(message.author, message.channel, qSleepTracker))) {
-    return false;
-  }
-  qSleepTracker.attachment = collected.attachments.size > 0 ? new Discord.Attachment(collected.attachments.first().url) : null;
-  qSleepTracker.answer = collected.content ? collected.content : "";
-  return true;
-}
-
-function getMember(message) {
-  const guild = getGuild(message);
-  const userId = message.author.id;
-  return guild.members.find(member => member.user.id === userId);
-}
-
-function getChannel(message, channelName) {
-  const guild = getGuild(message);
-  return guild.channels.find(channel => channel.name === channelName);
-}
-
-function getGuild(message) {
-  return message.client.guilds.first();
-}
-
-function hasRole(member, role) {
-  return member.roles.find(role => role.name == role);
 }
